+++ bike-banter/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@googlemaps/jest-mocks": "^2.22.6",
     "@supabase/supabase-js": "^2.49.4",
+    "@vis.gl/react-google-maps": "^1.5.4",
     "formdata-node": "^6.0.3",
     "next": "^15.3.5",
     "react": "^19.0.0",
@@ -21,21 +22,27 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
-    "@tailwindcss/postcss": "^4",
+    "@tailwindcss/cli": "^4.1.11",
+    "@tailwindcss/postcss": "^4.1.11",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
+    "@types/estree": "^1.0.8",
     "@types/jest": "^29.5.14",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "@types/ws": "^8.18.1",
+    "@typescript-eslint/types": "^8.38.0",
+    "autoprefixer": "^10.4.21",
     "dotenv": "^16.5.0",
     "eslint": "^9",
     "eslint-config-next": "15.3.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "postcss": "^8.5.6",
     "prettier": "3.5.3",
-    "tailwindcss": "^4",
+    "tailwindcss": "^4.1.11",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
   }
