--- bike-banter/src/app/__mocks__/google-maps-mock
@@ -1,16 +1,0 @@
-export const google = {
-  maps: {
-    Map: class {
-      constructor(element: HTMLElement, options: any) {}
-      setCenter() {}
-      setZoom() {}
-    },
-    Marker: class {
-      constructor(options: any) {}
-      setMap() {}
-    },
-    LatLng: class {
-      constructor(lat: number, lng: number) {}
-    },
-  },
-};
