+++ bike-banter/src/app/components/organisms/navbar-layout.tsx
@@ -1,11 +1,9 @@
-//fixed top-0 w-full z-50 h-16 bg-white shadow
 import Image from "next/image";
 
 export default function NavbarLayout() {
     return (
-        <nav className="flex h-[70px] items-center justify-between border-b border-gray-200 p-4">
+        <nav className="fixed top-0 left-0 w-full z-50 flex h-[70px] bg-white dark:bg-black text-black transition dark:text-white items-center justify-between border-b border-gray-200 p-4">
             <div className="flex items-center">
-                
                 <h1 className="flex items-center text-xl font-semibold capitalize"><Image
                     className=""
                     src="/bike-banter.svg"
@@ -15,22 +13,6 @@ export default function NavbarLayout() {
                     priority
                 />
                 </h1>
-            </div>
-            <div className="flex items-center space-x-2">
-                <a className="flex h-8 w-8 items-center justify-center rounded-full transition-colors hover:bg-gray-100" href="/f/inbox/new">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" className="lucide lucide-square-pen">
-                        <path d="M12 3H5a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7"></path>
-                        <path d="M18.375 2.625a1 1 0 0 1 3 3l-9.013 9.014a2 2 0 0 1-.853.505l-2.873.84a.5.5 0 0 1-.62-.62l.84-2.873a2 2 0 0 1 .506-.852z"></path>
-                    </svg>
-                </a>
-                <a className="flex h-8 w-8 items-center justify-center rounded-full transition-colors hover:bg-gray-100" href="/search">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" className="lucide lucide-search">
-                        <circle cx="11" cy="11" r="8">
-                        </circle>
-                        <path d="m21 21-4.3-4.3">
-                        </path>
-                    </svg>
-                </a>
             </div>
         </nav>
     );
