+++ bike-banter/src/app/page.tsx
@@ -2,8 +2,9 @@
 
 import Image from "next/image";
 import React, { useEffect } from 'react';
+//import './styles/globals.css';
 import { getUserLocation } from "@/utils/get-user-location";
-import UserGoogleMap  from "@/app/components/organisms/google-journey-map";
+import UserGoogleMap from "@/app/components/organisms/google-journey-map";
 import NavbarLayout from "@/app/components/organisms/navbar-layout";
 
 export default function Home() {
@@ -22,7 +23,7 @@ export default function Home() {
     };
 
     fetchUserLocation();
-  } , []);
+  }, []);
 
   return (
     <div className="grid grid-rows-[20px_1fr_20px] min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
@@ -39,7 +40,7 @@ export default function Home() {
             Ride
           </li>
         </ol>
-       <UserGoogleMap />
+        <UserGoogleMap />
         <div className="flex gap-4 items-center flex-col sm:flex-row">
           <a
             className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] font-medium text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:w-auto"
@@ -52,7 +53,7 @@ export default function Home() {
               src="/vercel.svg"
               alt="Vercel logomark"
               width={20}
-              height={20} 
+              height={20}
             />
             Deploy now
           </a>
@@ -61,7 +62,7 @@ export default function Home() {
             href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
             target="_blank"
             rel="noopener noreferrer"
-            >
+          >
             Read our docs
           </a>
         </div>
